+++ ass5_roll_test3.c
@@ -2,7 +2,7 @@ int test = 1;
 int main();
 int main () {
     int arr_int[20];
-    double arr_d[43];
+    float arr_d[43];
     int i=0;
     int n=9;
     for(i=0;i<n;i++)
